+++ Samples~/AssetWorkflowWithGUI/TestAssetSO.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (C) 2021-2024 Steffen Itterheim
+// Refer to included LICENSE file for terms and conditions.
+
+using UnityEngine;
+
+namespace CodeSmile_AssetDatabase_Demo
+{
+	// Just a placeholder
+	public class TestAssetSO : ScriptableObject {}
+}